Change Screen code
UI for all the screens
