+++ 2dFightTesting/2dFightTesting.csproj
@@ -49,6 +49,12 @@
     <Compile Include="BlueSamurai.cs" />
     <Compile Include="Character.cs" />
     <Compile Include="Attack.cs" />
+    <Compile Include="DemoScreen.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="DemoScreen.Designer.cs">
+      <DependentUpon>DemoScreen.cs</DependentUpon>
+    </Compile>
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -70,6 +76,9 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RedSamurai.cs" />
+    <EmbeddedResource Include="DemoScreen.resx">
+      <DependentUpon>DemoScreen.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
