+++ 2dFightTesting/DemoScreen.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace _2dFightTesting
+{
+    public partial class DemoScreen : UserControl
+    {
+        public DemoScreen()
+        {
+            InitializeComponent();
+        }
+
+        private void returnButton_Click(object sender, EventArgs e)
+        {
+            Form1.ChangeScreen(this, new MenuScreen());
+        }
+    }
+}
