+++ 2dFightTesting/MenuScreen.cs
@@ -19,7 +19,17 @@ namespace _2dFightTesting
 
         private void exitButton_Click(object sender, EventArgs e)
         {
+            Application.Exit();
+        }
 
+        private void playButton_Click(object sender, EventArgs e)
+        {
+            Form1.ChangeScreen(this, new GameScreen());
+        }
+
+        private void howToPlay_Click(object sender, EventArgs e)
+        {
+            Form1.ChangeScreen(this, new DemoScreen());
         }
     }
 }
